Detach DxLib initialiser from dxlib_spine_player.
